Major refactor of Sagemode tool class.
Major refactor for the Sagemode class, Sagemode is now modular and has been moved to it's own separate file. Sagemode class in Netwatch is now SagemodeMenu, and will only handle the transfer from Netwatch to the sagemode.py module.
