+++ .gitignore
@@ -166,7 +166,6 @@ src/accessories.py
 testRequirements.txt
 netwatchLogos.yml
 nmapHelpMenuData.json
-sagemode.py
 sagemode_old.py
 src/Nmap update utility.py
 src/colorTesting.py
